rename some consts
