+++ models/contacts.json
@@ -142,5 +142,17 @@
     "name": "Valentin",
     "email": "dui.in@egetlacus.ca",
     "phone": "(294) 840-6685"
+  },
+  {
+    "id": "e-YrJ-XLaalxFMKUh0YoJ",
+    "name": "Valentin",
+    "email": "dui.in@egetlacus.ca",
+    "phone": "(294) 840-6685"
+  },
+  {
+    "id": "_V52CHFTgETRjm2A6DE1Y",
+    "name": "Valentin",
+    "email": "dui.in@egetlacus.ca",
+    "phone": "(294) 840-6685"
   }
 ]
